% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hd.beta.R
\name{hd.beta}
\alias{hd.beta}
\title{Highest density interval for the beta distribution}
\usage{
hd.beta(c, d, rho = NULL, len = NULL, N = 1000)
}
\arguments{
\item{c}{First parameter of the beta distribution.}

\item{d}{Second parameter of the beta distribution.}

\item{rho}{A number in (0, 1) representing the fixed posterior probability of the HPD interval.}

\item{len}{A positive real number representing the fixed length for the HPD interval.}
}
\value{
Lower and upper bounds of the HD interval.
}
\description{
Highest density interval for the beta distribution
}
\examples{
.
}
