% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hd.diffbetas.R
\name{hd.diffbetas}
\alias{hd.diffbetas}
\title{Highest density interval for the difference of two beta distributions}
\usage{
hd.diffbetas(c1, d1, c2, d2, rho = NULL, len = NULL, N = 1000)
}
\arguments{
\item{c1}{}

\item{d1}{}

\item{c2}{}

\item{d2}{}

\item{rho}{}

\item{len}{}

\item{N}{}
}
\value{
Lower and upper bounds of the HD interval.
}
\description{
Highest density interval for the difference of two beta distributions
}
\examples{
.
}
