% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mss.bb.R
\name{mss.bb}
\alias{mss.bb}
\title{Minimum sample size for a single binomial parameter with a beta prior distribution}
\usage{
mss.bb(
  crit,
  c,
  d,
  rho.min = NULL,
  len = NULL,
  rho = NULL,
  len.max = NULL,
  R = 1000,
  n0 = 1
)
}
\arguments{
\item{crit}{A character string specifying the criterion. Criteria: "ACC", "ALC" and "ALCApprox".}

\item{c}{First parameter of the beta prior distribution.}

\item{d}{Second parameter of the beta prior distribution.}

\item{rho.min}{A number in (0, 1) representing the minimum admissible posterior probability for the HPD interval in the ACC.}

\item{len}{A positive real number representing the length of the HPD interval in the ACC.}

\item{rho}{A number in (0, 1) representing the posterior probability of the HPD in the ALC.}

\item{len.max}{A positive real number representing the maximum admissible length for the HPD interval in the ALC.}

\item{R}{Number of replicates used in the simulation. Default is 1000.}

\item{n0}{A positive integer representing the initial sample size in which the function will check the criterion. Default is 1.}
}
\value{
An integer representing the sample size.
}
\description{
Minimum sample size for a single binomial parameter with a beta prior distribution
}
\examples{
.
}
